apps/api/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "app" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "functional: Functional tests",
    "performance: Performance tests"
]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/api/html'
 --cov-report xml:'../../coverage/apps/api/coverage.xml'
 --html='../../reports/apps/api/unittests/html/index.html'
 --junitxml='../../reports/apps/api/unittests/junit.xml'
"""

[tool.poetry]
name = "api"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  # Исправляем директорию пакета
  [[tool.poetry.packages]]
  include = "app"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  flask = "^3.1.0"
  flask-cors = "^5.0.1"
  sqlalchemy = "^2.0.39"
  alembic = "^1.15.1"
  psycopg2-binary = "^2.9.10"
  ariadne = "^0.26.1"
  python-dotenv = "^1.1.0"
  gunicorn = "^23.0.0"
  graphql-core = "3.2.5"
  openpyxl = "^3.1.5"
  pandas = "^2.2.3"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.3.1"
  flake8 = "7.1.1"
  pytest = "8.3.4"
  pytest-sugar = "1.0.0"
  pytest-cov = "6.0.0"
  pytest-html = "4.1.1"
  pytest-xdist = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"